Zip File
Tables: The Librarian, Loan, Member, and Book tables were chosen to represent essential entities in the library system. Each table has specific attributes necessary for the management and operation of a library.
Data Types: The data types were chosen based on the nature of the data in each field. INT is used for numeric identifiers, VARCHAR for variable-length strings, and DATE for date fields, ensuring efficiency, flexibility, and accuracy in data storage.